--- libs/priors/priors/cosmologies.py
@@ -1,9 +1,0 @@
-import astropy.cosmology as cosmo
-
-
-def planck():
-    """
-    Return an `astropy` cosmology defined by the
-    Planck Collaboration 2015 paper
-    """
-    return cosmo.Planck15
