+++ projects/plots/plots/legacy/gwtc3.py
@@ -10,6 +10,8 @@ import scipy.stats as stats
 from astropy.utils.data import download_file
 from tqdm import tqdm
 
+from utils.cosmology import DEFAULT_COSMOLOGY
+
 catalog_results = {
     "GstLAL": {
         "Tb": 149.3747,
@@ -279,7 +281,8 @@ def main(
     sig_lognorm: float = 0.1,
     smax_ns: float = 0.4,
     smax_bh: float = 0.998,
-    cosmology: cosmo.Cosmology = cosmo.FlatwCDM(H0=67.9, Om0=0.3065, w0=-1),
+    # Note: original used cosmo.FlatwCDM(H0=67.9, Om0=0.3065, w0=-1),
+    cosmology: cosmo.Cosmology = DEFAULT_COSMOLOGY,
 ):
     known_pipelines = ["cwb", "gstlal", "mbta", "pycbc_bbh", "pycbc_hyperbank"]
     unknown_pipelines = set(pipelines) - set(known_pipelines)
