+++ projects/plots/plots/vizapp/pages/summary/sv.py
@@ -4,7 +4,6 @@ from typing import List, Optional
 
 import h5py
 import numpy as np
-from astropy.cosmology import Planck15 as cosmology
 from bilby.core.prior import PriorDict
 from bokeh.io import save
 from bokeh.layouts import gridplot
@@ -12,6 +11,7 @@ from bokeh.layouts import gridplot
 from ledger.events import EventSet, RecoveredInjectionSet
 from ledger.injections import InjectionParameterSet
 from priors.priors import log_normal_masses
+from utils.cosmology import DEFAULT_COSMOLOGY, get_astrophysical_volume
 
 from . import compute, utils
 from .gwtc3 import catalog_results
@@ -86,7 +86,7 @@ class SensitiveVolumePlot:
             decrange = None
         else:
             decrange = (decprior.minimum, decprior.maximum)
-        v0 = utils.get_astrophysical_volume(zmin, zmax, cosmology, decrange)
+        v0 = get_astrophysical_volume(zmin, zmax, DEFAULT_COSMOLOGY, decrange)
         v0 /= 10**9
         return v0
 
@@ -95,7 +95,7 @@ class SensitiveVolumePlot:
         for i, combo in enumerate(self.mass_combos):
             logging.info(f"Computing likelihoods under {combo} log normal")
             prior, _ = log_normal_masses(
-                *combo, sigma=self.sigma, cosmology=cosmology
+                *combo, sigma=self.sigma, cosmology=DEFAULT_COSMOLOGY
             )
             prob = get_prob(prior, self.foreground)
             rejected_prob = get_prob(prior, self.rejected_params)
